+++ gateway/dto/refresh.go
@@ -1,0 +1,6 @@
+package dto
+
+type RefreshResponse struct {
+	AccessToken  string `json:"access_token"`
+	RefreshToken string `json:"refresh_token"`
+}